+++ NEWS.md
@@ -77,6 +77,8 @@ Library improvements
   * `logspace` now accepts a `base` keyword argument to specify the base of the logarithmic
     range. The base defaults to 10 ([#22310]).
 
+  * Added `unique!` which is an inplace version of `unique` ([#20549]).
+
 Compiler/Runtime improvements
 -----------------------------